PostCSS implementation revised and improved. (#101)
- Added a method to get Monorepo Root Folder
- Using the new method to get Monorepo Root Folder
- Added a CollectionUtils.js method to ge Root Folder.
CollectionFactory.js now get Root Folder on its own using the new
method.
- Removed legacy file
- PostCSS now uses external config instead of Archie internal hardcoded
config.
- buildSnippets moved to SnippetBuilder.js, because duuuhhh! Also, it
removes a circular dependency.
- improves tests with dpdm, set it to fail if a circular dependency is
found and remove extra output.

+++ src/processors/StylesProcessor.js
@@ -1,6 +1,7 @@
 // Node JS internal imports
 import { unlink } from 'node:fs/promises'
 import path from 'path'
+import CollectionUtils from '../utils/CollectionUtils.js'
 // Archie Internal imports
 import PostCssProcessor from './PostCssProcessor.js'
 import SassPreProcessor from './SassPreProcessor.js'
@@ -18,27 +19,41 @@ class StylesProcessor {
   static async buildStyles (sourceFile, outputFile) {
     if (StylesUtils.isSassFile(sourceFile)) {
       return SassPreProcessor.processStyleSheet(sourceFile)
-    } else {
-      const css = await FileUtils.getFileContents(sourceFile)
-      return PostCssProcessor.processStyles(css, sourceFile, outputFile)
     }
+
+    const css = await FileUtils.getFileContents(sourceFile)
+
+    const postCssConfigFile = await CollectionUtils.findPostCssConfigFile()
+    if (await FileUtils.isReadable(postCssConfigFile)) {
+      const postCssConfig = await import(postCssConfigFile)
+      return PostCssProcessor.processStyles(css, sourceFile, outputFile, postCssConfig.default.plugins)
+    }
+
+    return css
   }
 
   /**
    * Create Styles Bundle
    * @param {string[]} stylesheets
    * @param {string} outputFile
+   * @param {string} [collectionName]
    * @return {Promise<string>}
    */
-  static async buildStylesBundle (stylesheets, outputFile) {
+  static async buildStylesBundle (stylesheets, outputFile, collectionName) {
     const masterStylesheet = path.join(os.tmpdir(), 'masterStylesheet.css')
     const masterStylesheetContents = StylesUtils.createMasterStylesheet(stylesheets)
 
     await FileUtils.writeFile(masterStylesheet, masterStylesheetContents)
 
-    const css = await PostCssProcessor.processStyles(masterStylesheetContents, masterStylesheet, outputFile)
-    await unlink(masterStylesheet)
-    return css
+    const postCssConfigFile = await CollectionUtils.findPostCssConfigFile(collectionName)
+    if (postCssConfigFile) {
+      const postCssConfig = await import(postCssConfigFile)
+      const css = await PostCssProcessor.processStyles(masterStylesheetContents, masterStylesheet, outputFile, postCssConfig.default)
+      await unlink(masterStylesheet)
+      return css
+    }
+
+    throw new Error('PostCSS config file not found')
   }
 }
